pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "shelfsight"
version = "0.1.0"
description = "Retail shelf detection pipeline (API + batch + AWS)"
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "python-multipart>=0.0.9",
  "Pillow>=10.0"
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "httpx>=0.27",
]

lint = [
  "ruff>=0.5",
  "black>=24.0",
]

dev = [
  "shelfsight[test,lint]",
]

aws = [
  "boto3>=1.28",
]

db = [
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9",
]

inference = [
  "ultralytics>=8.0",
]

data = [
  "pandas>=2.0",
]
# setuptools configuration for src-layout package discovery
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["shelfsight*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
ignore = ["B008"] # B008 conflicts with FastAPI dependency injection
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
# E/F: pycodestyle/pyflakes, I: isort, B: bugbear

[tool.black]
line-length = 100
target-version = ["py311"]